pyproject.toml
[tool.poetry]
name = "gsdss"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
smartsheet-python-sdk = "^2.105.1"
selenium = "^3.141.0"
twilio = "^6.50.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
